Ignore UID and PWD options from Excel
When the ODBC data source is used from Excel/Power Query, additional
options may be added to connection string automatically without showing
them to user. It was discovered in #58 that `UID` and `PWD` options are
added when user has credentials registered for this DSN in Excel.

DuckDB driver does not recognize these options and currently cannot
process them properly (see #59 for details). However the driver
knows about possible `trusted_connection` Excel/Power Query option and
ignores it successfully. It is proposed to also add `UID` and `PWD`
to the existing ignore list.

In the light of #60, that fixes the handling for unsupported options,
this change is not strictly necessary, but it still may be better to
consistently ignore all unneeded Excel/Power Query options and to not
register diagnostic messages for them.

Testing: existing Excel/Power Query ignore list test is extended to
include new options.

Fixes: #58