Fix SQLGetDiagRec for length-only calls
According to ODBC spec:

> If `MessageText` is `NULL`, `TextLengthPtr` will still return the
> total number of characters (excluding the null-termination character
> for character data) available to return in the buffer pointed to by
> `MessageText`.

Such length-only calls are actually used by the Windows ODBC Driver
Manager. Currently messages are returned in garbled form on Windows.

This change adds the logic to set the `TextLengthPtr` to the actual
message length when the `MessageText` is specified as `NULL`.

Note: the impl still can be improved to handle partial message reads
better (cover existing TODO there).

Testing: existing test is extended to cover this scenario.

Fixes: #56