Improve SQLGetInfo for MS Access
The ODBC client used in MS Access (also known as Microsoft Jet) has
specific requirements to the `SQLGetInfo` driver function. First, it
requires `SQL_ODBC_API_CONFORMANCE` to be supported. Second, it
requires the output parameter `StringLengthPtr` to be set by driver even
when the `InfoValuePtr` is set to non-character (integer) value.
Otherwise it fails with `Reserved error (-7701);` message and
disconnects immediately after the successfull connection.

This change adds a new helper `StoreWithLength` and applies it to all
non-character usages in `SQLGetInfo` and in `SQLGetConnectAttr`
functions. With it MS Access can establish the connection and start
running SQL queries. Though, it still fails on schema introspectio
(more Jet-specific fixes are needed in driver), so #39 remains open for
now.

Testing: length check for non-character data is added to the existing
test.