Fix varchar max column size in SQLGetTypeInfo
In the result set of `SQLGetTypeInfo` function the `COLUMN_SIZE`
column for `VACHAR` type must contain the length value in characters.

This change replaces `-1` value for this column with `MAX_INT` value.
This allows client tools based on Microsoft Power Query to use parameter
binding for `VARCHAR` columns/parameters.

Testing: existing tests, that were checking for `-1` value, are updated
to check for `MAX_INT`.

Fixes: #66