Fix SQLColumns output for decimal columns
`SQLColumns` was returning decimal column type as `SQL_DOUBLE` and also
was including type modifiers into the value in the `TYPE_NAME` column.

This confused PowerBI client which interpreted decimal columns as
binary columns and was unable to use them with parameters binding, and
this was breaking the query folding.

The patch removes type modifiers from the `TYPE_NAME` column and
changes the output type to `SQL_NUMERIC`. While `SQL_DECIMAL` may be
more appropriate for decimal columns, it appeared that `SQL_DECIMAL`
causes type conversion issues on PowerBI side, see details in the
issue linked in #77.

It also changes column types of the `SQLColumns` output to match the
types required by ODBC spec.

Testing: existing test is extended to cover decimal columns.

Fixes: #77