Support DATE and TIME bind parameter types
Support for `DATE` and `TIME` value was lacking in
`SQLBindParameter`/`SQLExecute`. This change adds such support the
same way as `TIMESTAMP` parameters are handled.

Besides that, when fetching records with `SQLGetData` there was no
conversion from DB `TIME` type to ODBC `TIMESTAMP` type. Such
conversion for some reason does not exist in the engine. But PyODBC uses
it by always fetching `TIME` columns as `TIMESTAMP`. This conversion is
added manually.

Testing: existing test adjusted when fetching `TIME` as `TIMESTAMP`, new
test is added to cover `SQLBindParameter` with `TIMESTAMP`, `DATE` and
`TIME` types.

Fixes: #91