Use push event to dispatch a rebuild after vendoring sources
Currently `Vendor.yml` workflows pushes updated sources to the repo and
then dispatches `ODBC.yml` workflow directly. The push on GitHub is not
synchronous, so there are no gaurantees that `ODBC.yml` will see the
pushed changes even if we wait for 10 seconds before the dispatch.

This change instead relies on a `push` event to start the `ODBC.yml`
workflow after the push. By default GitHub generates on-time token
`GITHUB_TOKEN` that is used as a credentials during the workflow run.
And this `GITHUB_TOKEN` is restricted to not trigger the `push` event.

Thus this change uses `secrets.PAT_TOKEN` token instead of the
`GITHUB_TOKEN`. This `PAT_TOKEN` needs to be added as a secret to this
repo. Minimal required permissions for it are `Contents: read-write`.

Testing: checked in another repo that this approach actually triggers
the push event.