Use PRs when importing engine sources
Currently `Vendor.yml` workflow imports DuckDB engine sources and
commits the changes to the `main` branch straight away.

This PR changes this to import engine changes into a `vendoring`
branch and create a PR on that branch. This PR triggers `ODBC.yml`
tests runs. Additional job is added to `ODBC.yml` to merge the PRs
on `vendoring` branch automatically after all tests are passed.

When the import happens and the PR on `vendoring` branch is still open
(when previous import did not pass the tests), then the imported
changes are added to the `vendoring` branch and the PR is closed and
re-opened to trigger the `ODBC.yml` workflow run.

Non-default personal access token (`PAT_TOKEN` secret) must be used for
creation or re-opening of the PR because the default token
`GITHUB_TOKEN` is restricted to not trigger `pull_request` event. All
other operations, besides the PR create/re-open, use the default token.

Merge is done by rebasing all `vendoring` commits onto `main` branch,
thus the history of the imports should look the same (linear) as it
looks now with direct pushes.

Testing: debugged this on a separate repo.