Update Ubuntu image used for clang-format runs
ODBC repo uses `clang-format-11` and `ubuntu-20.04` image to run format
checks. Ubuntu 20.04 is no longer supported on GitHub hosted runners:

https://github.com/actions/runner-images/issues/11101

As `ubuntu-latest` image does not include `clang-format-11` anymore,
the proposal is to move the image to `ubuntu-22.04` and then later
update `clang-format` to a newer version and move the image to
`ubuntu-latest`.