Use 4-part version numbers for releases
From time to time it may be necessary to make an updated release
keeping the engine version exactly the same as before and only updating
the ODBC part. It is proposed to use 4-part version number like
`v1.2.2.1` having the last number specific for ODBC-only updates.

Note: to have this change applicable to `1.2.2.x` releases we need to
add it to the `1.2` branch too.