Support Vendor.yml workflow for stable branches
`Vendor.yml` workflow can be called from the mainline
`NotifyExternalRepository.yml` workflow on the `main` branch or on one
of stable branches like `v1.2-histionicus`. `Vendor.yml` is called with
cURL call that triggers `workflow_dispatch` event in which the branch
name is a mandatory parameter. It is available to workflow in
`github.ref_name` context variable.

This change updates the check for changed engine files to be performed
on a correct branch instead of a hardcoded `main`.

It also adds the target branch suffix to `vendoring` branch that is used
for vendoring PRs.