Fix the crash on allowed_paths config option
Currently options `allowed_paths` and `allowed_directories` require
to have an initialized FS instance attached to the DB config. Otherwise
the process of checking file separators in the option value fails with
`Attempted to dereference unique_ptr that is NULL`.

This is tracked in the engine in duckdb/duckdb#17128 and is likely to
be fixed soon, but meanwhile it is proposed to add the fix to ODBC to
not crash the host process when one of these options is set in DSN
config.

Additionally the logic added to set `enable_external_access` option
after all other options (otherwise path-related options are not allowed
to be set) and to handle backslashes in windows paths in path-related
options.

Testing: new test added that checks that files can/cannot be read when
`allowed_paths` and `allowed_directories` options are used.