Add unicode entry points and data conversion
This change implements unicode ODBC entry points so the driver can be
recognized by the Driver Manager as a unicode-enabled driver.

The conversion of the input and output data is implemented using
lenient routines added in #70.

Testing: basic test coverage is added for all `*W` calls; new test is
added to check unicode conversion in `SQLBindParameter` and
`SQLGetData`; existing `pyodbc` Python test is enhanced to check unicode data
conversion with calls passing through the Driver Manager.

Fixes: #69