Correct type mappings for .NET numeric types
This PR correct the type mapping changing the ODBC type to
`SQL_DECIMAL` for the following DuckDB types:

 - `UBIGINT`
 - `HUGEINT`
 - `UHUGEINT`
 - `DECIMAL`

.NET CLS type for these DB types is now `System.Decimal`.

Testing: new test-suite is added using `dotnet` on Linux. Currently
it only covers the type mapping from `OdbcConnection` to .NET CLS types.
It overlaps with existing Windows `dotnet` test-suite, but it appeared
challenging to iterate over Windows `dotnet` due to long compile times
on Windows. New test-suite is confirmed to work on Ubuntu-22.04 and
Fedora. It is added to `ODBC Tests` CI run.

Fixes: #93