Support retrieving widechar data in parts
This change implements support for returning variable-length data to
client with multiple subsequent calls to `SQLGetData`. The ODBC spec
logic for buffer/write/return sizing is relatively straightforward, but
there is a number of corner cases. Also the logic for char, widechar
and binary data differs slightly (different null-termination handling).
This change unifies this logic for all variable-length types in one
place.

Testing: new test is added that covers retrieving data by parts for
char, widechar and binary data. PyODBC test suite is extended to check
the retrieval of overly long strings and also enabled on CI on Windows
(to check the retrieval through the Windows Driver Manager).

Fixes: #121