Allow query execution with unbound parameters
When a query is prepared, DuckDB, among other things, determines the
column types of the result set. In some cases it is not possible to
determine column types based on the query text alone, for example:

```
SELECT ?
```

In such cases the resulting column type will be different depending on
the parameter value provided, that can be aso different between the
executions of the same prepared query.

In such cases DuckDB does not create the query plan when query is
prepared and instead marks the statement with "unbound parameters"
marker.

In our ODBC impl there is a strict check that query preparation must
produce the plan and "unbound parameters" queries are rejected. The
"prepare" call is the same for both `SQLExecDirect` and
`SQLPrepare` + `SQLExecute` cases. Even when the client supplies
parameter values before calling `SQLExecDirect`, DuckDB does not have
API to pass these parameters to "prepare" call.

Proposed change removed the strict check for "unbound parameters"
queries allowing them to be executed. It corrects the logic around
Implementation Row Descriptor (IRD - description of result set columns)
records, so for "unbound parameters" queries these IRD records are kept
empty until the query is executed, and then re-filled after every
execution.

This change should not affect the logic for normal queries (where all
parameters are bound) so should not break any existing working
use-cases.

Testing: new test is added that covers unbound parameters case with
both `SQLExecDirect` and `SQLPrepare` + `SQLExecute` approaches.

Fixes: #119