Support catalog name in SQLTables
This change adds support for `CatalogName` parameter in `SQLTables`
call that was previously ignored. It is handled the same way as
`SchemaName` and `TableName` parameters.

Additionally it fixes special cases in `SQLTables` when only catalogs,
schemas or table types are listed.

Testing: existing test is updated to cover catalog filtering and all
special cases.

Fixes: #134