Support catalog name in SQLTables (1.3)
This is a backport of the PR #140 to `v1.3-ossivalis` stable branch.

This change adds support for `CatalogName` parameter in `SQLTables`
call that was previously ignored. It is handled the same way as
`SchemaName` and `TableName` parameters.

Additionally it fixes special cases in `SQLTables` when only catalogs,
schemas or table types are listed.

Testing: existing test is updated to cover catalog filtering and all
special cases.

Fixes: #134