Support DuckLake in connection string option
There is a usability gap with using DuckLake with ODBC, when the
connection must be opened to `:memory:` first and then the following
statement must be executed:

```
ATTACH 'ducklake:path/to/lake.db'
```

Executing this additional `ATTACH` step, when using ODBC from GUI tools
like Excel, is cumbersome and may require non-trivial configuration.

It is proposed to add special support for `ducklake:` prefix in the
database path option, that can be specified in DSN configuration or
directly in the connection string.

When `database=ducklake:path/to/lake.db` is specified, then it is
proposed to connect to `:memory:` DB instead and execute the `ATTACH`
command to the specified `ducklake:` path automatically.

Testing: test coverage pending (DuckLake extension not yet available for
`main` branch).