Increase column size for MSDASQL linked source
This change increases the max column size for `VARCHAR` and `VARBINARY`
fields reported in `SQLDescribeCol` (or `SQLColAttribute`) call.

Per ODBC spec this value is defined as:

> The defined or maximum column size in characters of the column

In DuckDB `VARCHAR` fields don't have the upper limit. We would like to
return `std::numeric_limits<SQLINTEGER>::max()` as max size, but it
appeared that `MSDASQL` provider cannot read strings when the declared
column size is over `8000`, that is the max size of a `VARCHAR`
(excluding `VARCHAR(MAX)`) in MSSQL.

Testing: tested manually with local MSSQL; automated tests are
problematic as `OPENROWSET` calls in MSSQL are disabled by default under
`Ad Hoc Distributed Queries` permission.

Fixes: #168