Update C-API exported symbols
This change updates the lists of exported C-API symbols from the latest
available `duckdb.h`. This is necessary to fix the build after the
breaking change to `duckdb.h` was introduced in duckdb/duckdb#18547.

PS: In future, it may be better to implement more robust handling for
exported symbols - to generate the CAPI export lists from `duckdb.h`.