Fix SQLNumResultCols for CALL, EXECUTE (1.4)
This is a backport of the PR #241 to `v1.4-andium` stable branch.

This PR fixes the `SQLNumResultCols` function making it to return the
actual number of columns for `CALL`, `EXECUTE` and `EXPLAIN` queries.
Without that some ODBC clients were unable to fetch the result set for
such queries.

`EXPLAIN` output contains 1 row with 2 columns. First column is a
constant `physical_plan` (or `logical_plan`) and the second column
contains actual plan details as a `VARCHAR`.

Testing: new test added.

Fixes: #200