Fix NumResultsCol for extension statements (1.4)
This is a backport of the PR #300 to `v1.4-andium` stable branch.

This change allows the statements of type `EXTENSION_STATEMENT` to
return results.

Testing: there seems to be no easy way to cover this with the test suite
as we don't want to depend on more extensions than now in tests. Checked
manually with `duckpgq`.

Fixes: #200