Enable ccache on CI on Linux and macOS (1.4)
This is a backport of the PR #316 to `v1.4-andium` stable branch.

This change enables saving/loading `ccache` caches on Linux and macOS.
DuckDB engine version is used as the cache key to speed up back-to-back
(like PR + merge) builds.

Windows ccache setup is currently unclear (naive approach with `ninja`
breaks the build) and may be added in future.