Enable ccache on Windows CI (1.4)
This is a backport of the PR #317 to `v1.4-andium` stable branch.

This PR adds support for `ccache` to Windows builds on CI and enables
caches save/load. The `ccache` usage logic is the same as with Linux/Mac
added in #316.

Windows build configuration is changed from MSBuild to NMake to be able
to use `CC='ccache cl'` env vars. Ninja cannot be used because it does
not support MSVC resources files.

`SystemDataODBC_tests` .NET tests are disabled because they require
MSBuild. They can be re-enabled in future, besides them .NET usage is
tested in Linux runs.