Add test coverage for empty queries
This test-only PR adds a test that executes "empty" queries (with no SQL
statements) - empty string, whitespaces, comments to check the
consistent behaviour for such queries.

The behaviour of the DuckDB driver, in general, matches the behaviour of
the PostgreSQL ODBC driver - the "empty" query execution completes
successfully, but an attempt to access the result set returns an error.